requirements.txt
tensorflow==2.11.0
pandas==1.5.1
yfinance==0.1.87
matplotlib==3.6.2
numpy==1.23.5
PyYAML==6.0
dask==2022.11.1
scikit-learn==1.1.3
kneed==0.8.1
tslearn==0.5.2
scipy==1.9.3